pyproject.toml
[tool.poetry]
name = "telegram"
version = "0.1.0"
description = ""
authors = ["Robert Solovev <robert.solovev.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^21.5"
fastapi = {extras = ["all"], version = "^0.114.0"}
sqlalchemy = "^2.0.34"
requests = "^2.32.3"
requests-html = "^0.10.0"
lxml-html-clean = "^0.2.2"
alembic = "^1.13.2"
python-dotenv = "^1.0.1"
celery = "^5.4.0"
redis = "3.5.3"
kombu = "^5.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
